scripts/requirements-bible.txt
httpx
beautifulsoup4
lxml
python-dotenv
supabase
